alteracao do bd para aceitar numero decimal nos valores do projeto do tipo 8

+++ app/database/migrations/2014_03_05_002059_create_project_monthly_progresses_table.php
@@ -20,7 +20,7 @@ class CreateProjectMonthlyProgressesTable extends Migration {
             $table->foreign('prefecture_id')->references('id')->on('prefectures')->onDelete('cascade');
             $table->integer('goal_target')->nullable();
             $table->datetime('month_year');
-            $table->integer('value')->nullable();
+            $table->decimal('value',15,2)->nullable();
             $table->timestamps();
         });
     }
